fix(logger): changed logging based on the item's type
log time was calculated when the item was processed not added; this PR corrects that and changes the signature for adding an item. also changes the context to always be an array

Refs [#15411]
