+++ tests/unit/Logger/Adapter/Noop/GetSetFormatterCest.php
