--- tests/unit/Logger/Adapter/Stream/GetFormatterCest.php
@@ -1,47 +1,0 @@
-<?php
-
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalcon.io>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Phalcon\Test\Unit\Logger\Adapter\Stream;
-
-use Phalcon\Logger\Adapter\Stream;
-use Phalcon\Logger\Formatter\FormatterInterface;
-use Phalcon\Logger\Formatter\Line;
-use UnitTester;
-
-class GetFormatterCest
-{
-    /**
-     * Tests Phalcon\Logger\Adapter\Stream :: getFormatter()
-     */
-    public function loggerAdapterStreamGetFormatter(UnitTester $I)
-    {
-        $I->wantToTest('Logger\Adapter\Stream - getFormatter()');
-
-        $fileName = $I->getNewFileName('log', 'log');
-        $fileName = logsDir($fileName);
-
-        $adapter = new Stream($fileName);
-
-        $adapter->getFormatter(
-            new Line()
-        );
-
-        $I->assertInstanceOf(
-            FormatterInterface::class,
-            $adapter->getFormatter()
-        );
-
-        $adapter->close();
-        $I->safeDeleteFile($fileName);
-    }
-}
