+++ tests/unit/Logger/Adapter/Stream/GetSetFormatterCest.php
