+++ tests/unit/Logger/Adapter/Syslog/GetSetFormatterCest.php
