--- tests/unit/Logger/Formatter/Json/SetDateFormatCest.php
@@ -1,39 +1,0 @@
-<?php
-
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalcon.io>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Phalcon\Test\Unit\Logger\Formatter\Json;
-
-use Phalcon\Logger\Formatter\Json;
-use UnitTester;
-
-class SetDateFormatCest
-{
-    /**
-     * Tests Phalcon\Logger\Formatter\Json :: setDateFormat()
-     *
-     * @author Phalcon Team <team@phalcon.io>
-     * @since  2018-11-13
-     */
-    public function loggerFormatterJsonSetDateFormat(UnitTester $I)
-    {
-        $I->wantToTest('Logger\Formatter\Json - setDateFormat()');
-
-        $formatter = new Json();
-        $format    = 'YmdHis';
-        $formatter->setDateFormat($format);
-
-        $actual = $formatter->getDateFormat();
-
-        $I->assertEquals($format, $actual);
-    }
-}
